Add `eig3volume.mexw64`

+++ code/dependencies/FRANGI_functions/imgaussian.m
@@ -30,7 +30,7 @@ if(sigma>0)
     % Make 1D Gaussian kernel
     x=-ceil(siz/2):ceil(siz/2);
     H = exp(-(x.^2/(2*sigma^2)));
-    H = H/sum(H(:));
+    H = double(H/sum(H(:)));
 
     % Filter each dimension with the 1D Gaussian kernels\
     if(ndims(I)==1)
